Added JsonIgnore Attribute
When JsonIgnore is applied to a field or property, that member is
ignored when deserializing and/or serializing.
You can specify JsonIgnoreWhen.Serializing to only ignore
serialization, *.Deserializing to only ignore deserializing, or default
which is to ignore the field either way.
Added unit tests for JsonIgnore, mroe to come.
Updated Makefile.
Updated .gitignore to include debug and OSX files.